Working Parser for XML files to read tissue parameters
xml files of the form in the oflder Testing/TestData can be read into
a struct containing data on the parameters of different tissue types.

this way easily new parameters can be added if needed.
The parser just returns a vector of all the possible tissue types.
A phantom as a container with integer numbers corresponding to different
tissue types can now point to the entry in this list and access its
numerical content in order to generate artificial contrast.