Remove 'Size' method from ImageData
I put this in so that the workspace of Matlab would show the number of dimensions of the image. For example, it would say `64x64x64 ImageData`. However, by doing this, I could no longer distinguish a single image from an array of images. Is `1x2 ImageData` 2 image datas, or 1 image data with 1x2 voxels? Would be good to find a way to distinguish the two so that I can put this functionality back.