Fix error in c implementation of get_data
Need to return as type `float`. So create deep_copy of original and cast to `float`. But then I was copying data from the original as opposed to the copy that had been casted to the correct datatype. This means it only worked, if the original type was float (which is why I didn't see the problem). Fixed now.