Use target_include_directories
Following from #250, if `target_include_directories` is used, then the header files will be available when a library is linked.

Some `.h` files are still at the same level as the `.cpp` files, whereas the majority are in `include/sirf/engine`. This means that for the time being, we need to include to directories per library.