Resampling complex images in python and c++ (#558)
The resampling of complex images. Internally, in the resampler, images are converted from the `ImageData` (SIRF base class) into two `NiftiImageData` for the real and imaginary components. If the input image is real, then the imaginary component will remain a null pointer. We then resample each of the real and imaginary components and then join the output together at the end. The two components of the complex `NiftiImageData` are stored in an internal class called `ComplexNiftiImageData`.

* Also required modification of the `NiftiImageData` assignment/copy operators. Now, they all call the same function, regardless of whether the `rhs` argument was of type `NiftiImageData` or `ImageData` (SIRF base class).