SPM functionality (#540)
Wrap `spm_realign` and `spm_reslice` functionality from SPM. Since SPM is written in Matlab, the SPM functionality will only be compiled if Matlab is present.

The SIRF C++ code calls the Matlab engine API, which calls SPM. This therefore means that we can wrap to Python and Matlab. The result is that in SIRF, SPM functionality is available from C++, Matlab and Python.

As part of this PR, registration of multiple floating images is now provided for algorithms that support it (currently only SPM). We now use `registration->get_output(i)` to get the `i`ith registered image, and similarly for the `i`th transformation matrix, deformation field, etc. if `registration->get_output()` is used with no argument, the first will be returned to ensure backwards compatibility.

Lastly, a middle (abstract) class was added to the registration hierarchy. The structure is now as follows:

- `Registration` 
    - `NiftiBasedRegistration` 
        - `NiftyRegistration`
            - `Aladin`
            - `F3d`
        - `SPMRegistration`