consistent use of std::shared_ptr vs stir::shared_ptr
We now use std::shared_ptr for SIRF objects and stir::shared_ptr for
STIR objects. Normally they are the same, but that isn't guaranteed.

In the previous code, there was a mix of usages, which was never tested
as SIRF was only ever compiled with stir::shared_ptr=std::shared_ptr.