CMake fixes for dependencies
- add more conditionals to see if certain things can be built
- split csirf library into 2: csirf is now the C++ library, while
cinterface-sirf contains the C wrappers.
This is probably more logical, but also avoids a circular dependency on
ImageDataWrapper (which sits in csyn, but depends on creg etc).
ImageDataWrapper is currently not using in csirf, but only in cinterface-sirf
(i.e. Python/MATLAB)
- remove the option DISABLE_Synergistic when building Python/MATLAB stuff as
disabling it would lead to linking errors (as they need ImageDataWrapper)