Managed to switch to read-all-acquisitions default for MR AcquisitionData
Failures cased by noise calibration data forced ignoring it, which was
a poor solution. It turned out that ignoring these data just on the
preliminary test run in AcquisitionProcessor chain (to find out whether
the chain needs AcquisitionFinishGadget) appears to fix the problem
save for grappa2_1rep.h5 an grappa2_6rep.h5 data. Possible reason for
failures with the latter data is the absence of the measurementInformation
in xml headers. Assuming this is true, we can now use the proper
read-all-acquisition default when reading MR acquisition data, fixing
grappa2 data to be addressed by a separate issue.